refactor(src): split finalize/mod.rs by responsibility
